+++ VIStk/Widgets/__init__.py
@@ -15,6 +15,10 @@ from VIStk.Widgets._AutocompleteEntry import AutocompleteEntry
 from VIStk.Widgets._DateEntry import DateEntry
 from VIStk.Widgets._ContextMenu import ContextMenu
 from VIStk.Widgets._SettingsWindow import SettingsWindow
+from VIStk.Widgets._vWidget import vWidget, make_rounded_image
+from VIStk.Widgets._vLabel import vLabel
+from VIStk.Widgets._vButton import vButton
+from VIStk.Widgets._vFrame import vFrame
 
 
 __all__ = ["VISMenu",
@@ -35,4 +39,9 @@ __all__ = ["VISMenu",
            "DateEntry",
            "ContextMenu",
            "SettingsWindow",
+           "vWidget",
+           "make_rounded_image",
+           "vLabel",
+           "vButton",
+           "vFrame",
            ]